Create README.md
This application was to show the functionality of JavaScript. The html was intentionally left very simple to show off some of the things that were accomplished using Javascript.
Java, JavaScript, Spring, Thymeleaf, and Hibernate were all used in this project. This app is full stack, which starts at JavaScript and communicates with the Database.